requirements.txt
aioredis==1.3.1
arabic-reshaper==2.1.3  
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.1.8
django-bootstrap-breadcrumbs==0.9.2
django-cors-headers==3.8.0
django-filter==2.4.0
django-tables2==2.3.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
future==0.18.2
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
msgpack==1.0.2
openpyxl==3.0.7
Pillow==8.2.0
#psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.1
qrcode==6.1
reportlab==3.6.6
requests==2.28.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
#twisted-iocpsupport==1.0.1
txaio==21.2.1
webencodings==0.5.1
#uWSGI==2.0.19.1
xhtml2pdf==0.2.5
zope.interface==5.4.0
uvicorn==0.13.4
